Split quilts tests from publish. Also use --all-features instead of full
